Add BRAM register to axis_fifo_xilinx_bram and ram_dual_port_2clk
This adds a parameter to enable the output register of the BRAM
defined in ram_dual_port_2clk. This is used by default in the
AXI-S BRAM FIFO to improve timing.

The read logic in axis_fifo_xilinx_bram has been rewritten because
it didn't seem to be easy to generalize to a read latency greater
than 1 cycle.

Additionally, some SystemVerilog coding style improvements have
been done in the two modified modules.