mux: fix bug caused by single-beat packets in port 0
If the mux is in IDLE state, its internal minimal_fifo is not full
and a single-beat packet comes in port 0, the mux would remain
in state SELECT0 instead of IDLE after forwarding this packet.
This would prevent packets in other ports from being forwarded
until a new packet comes in port 0.

The cause of the bug is that in IDLE, the mux acts similarly to
the SELECT0 state: the input 0 is connected to the minimal_fifo and
the other inputs have TREADY set to 0. Therefore, a single-beat
packet in input 0 is completely processed in a single cycle in the
IDLE state if the minimal_fifo is not full. The state machine needs
to take this into account and remain in the IDLE state instead of
going to SELECT0 when this happens.

The problem only happens in the first input, because if a packet
comes in one of the other inputs (say input 1), first a cycle needs
to be spent going from IDLE to SELECT1, and then from SELECT1 the
state machine will go to IDLE whenever TLAST happens.

This commit fixes this bug in axis_mux2, axis_mux4 and axis_mux8.